% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{okf_available}
\alias{okf_available}
\title{Check if okf binary is available}
\usage{
okf_available(verbose = FALSE)
}
\arguments{
\item{verbose}{Print diagnostic info}
}
\value{
TRUE if binary found
}
\description{
Check if okf binary is available
}
