% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okf.R
\name{okf_bin}
\alias{okf_bin}
\title{Find the okf binary bundled with the package}
\usage{
okf_bin(verbose = FALSE)
}
\arguments{
\item{verbose}{Print diagnostic info}
}
\value{
Path to the okf binary for current platform
}
\description{
Find the okf binary bundled with the package
}
\keyword{internal}
